Moved Ball Spawn for each player

+++ ChitPhysicsTest1/Assets/Scripts/SpawnPosition.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class SpawnPosition : MonoBehaviour
 {
     public bool IsOccupied = false;
-    public int occupyingClientId;
+    public ulong occupyingClientId;
 }
